Feat: Adiciona Mapper e DTO, ajusta Service e Controller para a Entidade Votacao

+++ src/main/java/com/example/Deputados/Votacao/VotacaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface VotacaoRepository extends JpaRepository<VotacaoModel, Integer>{
+public interface VotacaoRepository extends JpaRepository<VotacaoModel, Long>{
     
 }
